(misc) move registration to providers package
