+++ providers/registration/file_content.go
