+++ providers/registration/testdata/sample.json
